Move specific code to specific places
