+++ src/operations/search/aggregations/mod.rs
@@ -16,6 +16,9 @@
 
 //! Implementation of ElasticSearch [aggregations](https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations.html)
 
+pub mod metrics;
+pub mod bucket;
+
 use std::collections::{BTreeMap, HashMap};
 use std::marker::PhantomData;
 
@@ -95,16 +98,6 @@ impl<'a> Script<'a> {
     }
 }
 
-macro_rules! metrics_agg {
-    ($b:ident) => {
-        impl<'a> From<$b<'a>> for Aggregation<'a> {
-            fn from(from: $b<'a>) -> Aggregation<'a> {
-                Aggregation::Metrics(MetricsAggregation::$b(from))
-            }
-        }
-    }
-}
-
 /// A common pattern is for an aggregation to accept a field or a script
 #[derive(Debug)]
 pub enum FieldOrScript<'a> {
@@ -181,1105 +174,15 @@ macro_rules! field_or_script_to_json {
     }
 }
 
-/// Min aggregation
-#[derive(Debug, Serialize)]
-pub struct Min<'a>(FieldOrScript<'a>);
-
-field_or_script_new!(Min);
-field_or_script_to_json!(Min);
-metrics_agg!(Min);
-
-// /// Max aggregation
-// #[derive(Debug)]
-// pub struct Max<'a>(FieldOrScript<'a>);
-
-// field_or_script_new!(Max);
-// field_or_script_to_json!(Max);
-// metrics_agg!(Max);
-
-// /// Sum aggregation
-// #[derive(Debug)]
-// pub struct Sum<'a>(FieldOrScript<'a>);
-
-// field_or_script_new!(Sum);
-// field_or_script_to_json!(Sum);
-// metrics_agg!(Sum);
-
-// /// Avg aggregation
-// #[derive(Debug)]
-// pub struct Avg<'a>(FieldOrScript<'a>);
-
-// field_or_script_new!(Avg);
-// field_or_script_to_json!(Avg);
-// metrics_agg!(Avg);
-
-// /// Stats aggregation
-// #[derive(Debug)]
-// pub struct Stats<'a>(FieldOrScript<'a>);
-
-// field_or_script_new!(Stats);
-// field_or_script_to_json!(Stats);
-// metrics_agg!(Stats);
-
-// /// Extended stats aggregation
-// #[derive(Debug)]
-// pub struct ExtendedStats<'a>(FieldOrScript<'a>);
-
-// field_or_script_new!(ExtendedStats);
-// field_or_script_to_json!(ExtendedStats);
-// metrics_agg!(ExtendedStats);
-
-// /// Value count aggregation
-// #[derive(Debug)]
-// pub struct ValueCount<'a>(FieldOrScript<'a>);
-
-// field_or_script_new!(ValueCount);
-// field_or_script_to_json!(ValueCount);
-// metrics_agg!(ValueCount);
-
-// /// Percentiles aggregation, see: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-percentile-aggregation.html
-// ///
-// /// # Examples
-// ///
-// /// ```
-// /// use rs_es::operations::search::aggregations::Percentiles;
-// ///
-// /// let p1 = Percentiles::new("field_name").with_compression(100u64);
-// /// let p2 = Percentiles::new("field_name").with_percents(vec![10.0, 20.0]);
-// /// ```
-// #[derive(Debug)]
-// pub struct Percentiles<'a> {
-//     fos:         FieldOrScript<'a>,
-//     percents:    Option<Vec<f64>>,
-//     compression: Option<u64>
-// }
-
-// impl<'a> Percentiles<'a> {
-//     pub fn new<F: Into<FieldOrScript<'a>>>(fos: F) -> Percentiles<'a> {
-//         Percentiles {
-//             fos:         fos.into(),
-//             percents:    None,
-//             compression: None
-//         }
-//     }
-
-//     add_field!(with_percents, percents, Vec<f64>);
-//     add_field!(with_compression, compression, u64);
-// }
-
-// impl<'a> ToJson for Percentiles<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         self.fos.add_to_object(&mut d);
-//         optional_add!(self, d, percents);
-//         optional_add!(self, d, compression);
-//         Json::Object(d)
-//     }
-// }
-
-// metrics_agg!(Percentiles);
-
-// /// Percentile Ranks aggregation
-// #[derive(Debug)]
-// pub struct PercentileRanks<'a> {
-//     fos:    FieldOrScript<'a>,
-//     values: Vec<f64>
-// }
-
-// impl<'a> PercentileRanks<'a> {
-//     pub fn new<F: Into<FieldOrScript<'a>>>(fos: F, vals: Vec<f64>) -> PercentileRanks<'a> {
-//         PercentileRanks {
-//             fos:    fos.into(),
-//             values: vals
-//         }
-//     }
-// }
-
-// impl<'a> ToJson for PercentileRanks<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         self.fos.add_to_object(&mut d);
-//         d.insert("values".to_owned(), self.values.to_json());
-//         Json::Object(d)
-//     }
-// }
-
-// metrics_agg!(PercentileRanks);
-
-// /// Cardinality aggregation
-// #[derive(Debug)]
-// pub struct Cardinality<'a> {
-//     fos:                 FieldOrScript<'a>,
-//     precision_threshold: Option<u64>,
-//     rehash:              Option<bool>
-// }
-
-// impl<'a> Cardinality<'a> {
-//     pub fn new<F: Into<FieldOrScript<'a>>>(fos: F) -> Cardinality<'a> {
-//         Cardinality {
-//             fos:                 fos.into(),
-//             precision_threshold: None,
-//             rehash:              None
-//         }
-//     }
-
-//     add_field!(with_precision_threshold, precision_threshold, u64);
-//     add_field!(with_rehash, rehash, bool);
-// }
-
-// impl<'a> ToJson for Cardinality<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         self.fos.add_to_object(&mut d);
-//         optional_add!(self, d, precision_threshold);
-//         optional_add!(self, d, rehash);
-//         Json::Object(d)
-//     }
-// }
-
-// metrics_agg!(Cardinality);
-
-// /// Geo Bounds aggregation
-// #[derive(Debug)]
-// pub struct GeoBounds<'a> {
-//     field:          &'a str,
-//     wrap_longitude: Option<bool>
-// }
-
-// impl<'a> GeoBounds<'a> {
-//     pub fn new(field: &'a str) -> GeoBounds {
-//         GeoBounds {
-//             field: field,
-//             wrap_longitude: None
-//         }
-//     }
-
-//     add_field!(with_wrap_longitude, wrap_longitude, bool);
-// }
-
-// impl<'a> ToJson for GeoBounds<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         d.insert("field".to_owned(), self.field.to_json());
-//         optional_add!(self, d, wrap_longitude);
-//         Json::Object(d)
-//     }
-// }
-
-// metrics_agg!(GeoBounds);
-
-/// Scripted method aggregation
-#[derive(Debug)]
-pub struct ScriptedMetric<'a> {
-    init_script:         Option<&'a str>,
-    map_script:          &'a str,
-    combine_script:      Option<&'a str>,
-    reduce_script:       Option<&'a str>,
-    params:              Option<Json>,
-    reduce_params:       Option<Json>,
-    lang:                Option<&'a str>,
-    init_script_file:    Option<&'a str>,
-    init_script_id:      Option<&'a str>,
-    map_script_file:     Option<&'a str>,
-    map_script_id:       Option<&'a str>,
-    combine_script_file: Option<&'a str>,
-    combine_script_id:   Option<&'a str>,
-    reduce_script_file:  Option<&'a str>,
-    reduce_script_id:    Option<&'a str>
-}
-
-impl<'a> ScriptedMetric<'a> {
-    pub fn new(map_script: &'a str) -> ScriptedMetric<'a> {
-        ScriptedMetric {
-            init_script:         None,
-            map_script:          map_script,
-            combine_script:      None,
-            reduce_script:       None,
-            params:              None,
-            reduce_params:       None,
-            lang:                None,
-            init_script_file:    None,
-            init_script_id:      None,
-            map_script_file:     None,
-            map_script_id:       None,
-            combine_script_file: None,
-            combine_script_id:   None,
-            reduce_script_file:  None,
-            reduce_script_id:    None
-        }
-    }
-
-    add_field!(with_init_script, init_script, &'a str);
-    add_field!(with_combine_script, combine_script, &'a str);
-    add_field!(with_reduce_script, reduce_script, &'a str);
-    add_field!(with_params, params, Json);
-    add_field!(with_reduce_params, reduce_params, Json);
-    add_field!(with_lang, lang, &'a str);
-    add_field!(with_init_script_file, init_script_file, &'a str);
-    add_field!(with_init_script_id, init_script_id, &'a str);
-    add_field!(with_map_script_file, map_script_file, &'a str);
-    add_field!(with_map_script_id, map_script_id, &'a str);
-    add_field!(with_combine_script_file, combine_script_file, &'a str);
-    add_field!(with_combine_script_id, combine_script_id, &'a str);
-    add_field!(with_reduce_script_file, reduce_script_file, &'a str);
-    add_field!(with_reduce_script_id, reduce_script_id, &'a str);
-}
-
-impl<'a> ToJson for ScriptedMetric<'a> {
-    fn to_json(&self) -> Json {
-        let mut d = BTreeMap::new();
-        d.insert("map_script".to_owned(), self.map_script.to_json());
-        optional_add!(self, d, init_script);
-        optional_add!(self, d, combine_script);
-        optional_add!(self, d, reduce_script);
-        optional_add!(self, d, params);
-        optional_add!(self, d, reduce_params);
-        optional_add!(self, d, lang);
-        optional_add!(self, d, init_script_file);
-        optional_add!(self, d, init_script_id);
-        optional_add!(self, d, map_script_file);
-        optional_add!(self, d, map_script_id);
-        optional_add!(self, d, combine_script_file);
-        optional_add!(self, d, combine_script_id);
-        optional_add!(self, d, reduce_script_file);
-        optional_add!(self, d, reduce_script_id);
-        Json::Object(d)
-    }
-}
-
-/// Individual aggregations and their options
-#[derive(Debug, Serialize)]
-pub enum MetricsAggregation<'a> {
-    #[serde(rename="min")]
-    Min(Min<'a>),
-    // Max(Max<'a>),
-    // Sum(Sum<'a>),
-    // Avg(Avg<'a>),
-    // Stats(Stats<'a>),
-    // ExtendedStats(ExtendedStats<'a>),
-    // ValueCount(ValueCount<'a>),
-    // Percentiles(Percentiles<'a>),
-    // PercentileRanks(PercentileRanks<'a>),
-    // Cardinality(Cardinality<'a>),
-    // GeoBounds(GeoBounds<'a>),
-    // ScriptedMetric(ScriptedMetric<'a>)
-}
-
-// TODO deprecated
-// impl<'a> ToJson for MetricsAggregation<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         match self {
-//             &MetricsAggregation::Min(ref min_agg) => {
-//                 d.insert("min".to_owned(), min_agg.to_json());
-//             },
-//             &MetricsAggregation::Max(ref max_agg) => {
-//                 d.insert("max".to_owned(), max_agg.to_json());
-//             },
-//             &MetricsAggregation::Sum(ref sum_agg) => {
-//                 d.insert("sum".to_owned(), sum_agg.to_json());
-//             },
-//             &MetricsAggregation::Avg(ref avg_agg) => {
-//                 d.insert("avg".to_owned(), avg_agg.to_json());
-//             },
-//             &MetricsAggregation::Stats(ref stats_agg) => {
-//                 d.insert("stats".to_owned(), stats_agg.to_json());
-//             },
-//             &MetricsAggregation::ExtendedStats(ref ext_stat_agg) => {
-//                 d.insert("extended_stats".to_owned(), ext_stat_agg.to_json());
-//             },
-//             &MetricsAggregation::ValueCount(ref vc_agg) => {
-//                 d.insert("value_count".to_owned(), vc_agg.to_json());
-//             },
-//             &MetricsAggregation::Percentiles(ref pc_agg) => {
-//                 d.insert("percentiles".to_owned(), pc_agg.to_json());
-//             },
-//             &MetricsAggregation::PercentileRanks(ref pr_agg) => {
-//                 d.insert("percentile_ranks".to_owned(), pr_agg.to_json());
-//             },
-//             &MetricsAggregation::Cardinality(ref card_agg) => {
-//                 d.insert("cardinality".to_owned(), card_agg.to_json());
-//             },
-//             &MetricsAggregation::GeoBounds(ref gb_agg) => {
-//                 d.insert("geo_bounds".to_owned(), gb_agg.to_json());
-//             },
-//             &MetricsAggregation::ScriptedMetric(ref sm_agg) => {
-//                 d.insert("scripted_metric".to_owned(), sm_agg.to_json());
-//             }
-//         }
-//         Json::Object(d)
-//     }
-// }
-
-// Bucket aggregations
-
-macro_rules! bucket_agg {
-    ($b:ident) => {
-        impl<'a> From<($b<'a>, Aggregations<'a>)> for Aggregation<'a> {
-            fn from(from: ($b<'a>, Aggregations<'a>)) -> Aggregation<'a> {
-                Aggregation::Bucket(BucketAggregation::$b(from.0), Some(from.1))
-            }
-        }
-
-        impl<'a> From<$b<'a>> for Aggregation<'a> {
-            fn from(from: $b<'a>) -> Aggregation<'a> {
-                Aggregation::Bucket(BucketAggregation::$b(from), None)
-            }
-        }
-    }
-}
-
-/// Global aggregation, defines a single global bucket.  Can only be used as a
-/// top-level aggregation.  See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-global-aggregation.html
-#[derive(Debug, Serialize)]
-pub struct Global<'a> {
-    /// Needed for lifecycle reasons
-    phantom: PhantomData<&'a str>
-}
-
-impl<'a> Global<'a> {
-    pub fn new() -> Global<'a> {
-        Global {
-            phantom: PhantomData
-        }
-    }
-}
-
-// TODO - deprecated
-impl<'a> ToJson for Global<'a> {
-    fn to_json(&self) -> Json {
-        Json::Object(BTreeMap::new())
-    }
-}
-
-bucket_agg!(Global);
-
-// /// Filter aggregation
-// #[derive(Debug)]
-// pub struct Filter<'a> {
-//     // TODO - Query is now an enum with a `Box` it might be simpler, with little
-//     // side-effects to own it instead
-//     filter: &'a query::Query
-// }
-
-// impl<'a> Filter<'a> {
-//     pub fn new(filter: &'a query::Query) -> Filter<'a> {
-//         Filter {
-//             filter: filter
-//         }
-//     }
-// }
-
-// impl<'a> ToJson for Filter<'a> {
-//     fn to_json(&self) -> Json {
-//         self.filter.to_json()
-//     }
-// }
-
-// bucket_agg!(Filter);
-
-// /// Filters aggregation
-// #[derive(Debug)]
-// pub struct Filters<'a> {
-//     filters: HashMap<&'a str, &'a query::Query>
-// }
-
-// impl<'a> Filters<'a> {
-//     pub fn new(filters: HashMap<&'a str, &'a query::Query>) -> Filters<'a> {
-//         Filters {
-//             filters: filters
-//         }
-//     }
-// }
-
-// impl<'a> From<Vec<(&'a str, &'a query::Query)>> for Filters<'a> {
-//     fn from(from: Vec<(&'a str, &'a query::Query)>) -> Filters<'a> {
-//         let mut filters = HashMap::with_capacity(from.len());
-//         for (k, v) in from {
-//             filters.insert(k, v);
-//         }
-//         Filters::new(filters)
-//     }
-// }
-
-// impl<'a> ToJson for Filters<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         let mut inner = BTreeMap::new();
-//         for (&k, v) in self.filters.iter() {
-//             inner.insert(k.to_owned(), v.to_json());
-//         }
-//         d.insert("filters".to_owned(), Json::Object(inner));
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(Filters);
-
-// /// Missing aggregation
-// #[derive(Debug)]
-// pub struct Missing<'a> {
-//     pub field: &'a str
-// }
-
-// impl<'a> Missing<'a> {
-//     pub fn new(field: &'a str) -> Missing<'a> {
-//         Missing {
-//             field: field
-//         }
-//     }
-// }
-
-// impl<'a> ToJson for Missing<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         d.insert("field".to_owned(), self.field.to_json());
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(Missing);
-
-// /// Nested aggregation
-// #[derive(Debug)]
-// pub struct Nested<'a> {
-//     pub path: &'a str
-// }
-
-// impl<'a> Nested<'a> {
-//     pub fn new(path: &'a str) -> Nested<'a> {
-//         Nested {
-//             path: path
-//         }
-//     }
-// }
-
-// impl<'a> ToJson for Nested<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         d.insert("path".to_owned(), self.path.to_json());
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(Nested);
-
-// /// Reverse nested aggregation, will produce an error if used anywhere other than
-// /// inside a nested aggregation.
-// ///
-// /// See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-reverse-nested-aggregation.html
-// #[derive(Debug)]
-// pub struct ReverseNested<'a> {
-//     /// Needed for lifecycle reasons
-//     phantom: PhantomData<&'a str>
-// }
-
-// impl<'a> ReverseNested<'a> {
-//     pub fn new() -> ReverseNested<'a> {
-//         ReverseNested {
-//             phantom: PhantomData
-//         }
-//     }
-// }
-
-// impl<'a> ToJson for ReverseNested<'a> {
-//     fn to_json(&self) -> Json {
-//         Json::Object(BTreeMap::new())
-//     }
-// }
-
-// bucket_agg!(ReverseNested);
-
-// /// Children aggregation - sub-aggregations run against the child document
-// #[derive(Debug)]
-// pub struct Children<'a> {
-//     doc_type: &'a str
-// }
-
-// impl<'a> Children<'a> {
-//     pub fn new(doc_type: &'a str) -> Children<'a> {
-//         Children {
-//             doc_type: doc_type
-//         }
-//     }
-// }
-
-// impl<'a> ToJson for Children<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         d.insert("type".to_owned(), self.doc_type.to_json());
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(Children);
-
-/// Order - used for some bucketing aggregations to determine the order of
-/// buckets
-#[derive(Debug)]
-pub enum OrderKey<'a> {
-    Count,
-    Key,
-    Term,
-    Expr(&'a str)
-}
-
-impl<'a> From<&'a str> for OrderKey<'a> {
-    fn from(from: &'a str) -> OrderKey<'a> {
-        OrderKey::Expr(from)
-    }
-}
-
-impl<'a> ToString for OrderKey<'a> {
-    fn to_string(&self) -> String {
-        match *self {
-            OrderKey::Count   => "_count".to_owned(),
-            OrderKey::Key     => "_key".to_owned(),
-            OrderKey::Term    => "_term".to_owned(),
-            OrderKey::Expr(e) => e.to_owned()
-        }
-    }
-}
-
-/// Used to define the ordering of buckets in a some bucketted aggregations
-///
-/// # Examples
-///
-/// ```
-/// use rs_es::operations::search::aggregations::{Order, OrderKey};
-///
-/// let order1 = Order::asc(OrderKey::Count);
-/// let order2 = Order::desc("field_name");
-/// ```
-///
-/// The first will produce a JSON fragment: `{"_count": "asc"}`; the second will
-/// produce a JSON fragment: `{"field_name", "desc"}`
-#[derive(Debug)]
-pub struct Order<'a>(OrderKey<'a>, super::Order);
-
-impl<'a> Order<'a> {
-    /// Create an `Order` ascending
-    pub fn asc<O: Into<OrderKey<'a>>>(key: O) -> Order<'a> {
-        Order(key.into(), super::Order::Asc)
-    }
-
-    /// Create an `Order` descending
-    pub fn desc<O: Into<OrderKey<'a>>>(key: O) -> Order<'a> {
-        Order(key.into(), super::Order::Desc)
-    }
-}
-
-impl<'a> ToJson for Order<'a> {
-    fn to_json(&self) -> Json {
-        let mut d = BTreeMap::new();
-        d.insert(self.0.to_string(), self.1.to_json());
-        Json::Object(d)
-    }
-}
-
-// /// Terms aggregation
-// #[derive(Debug)]
-// pub struct Terms<'a> {
-//     field:      FieldOrScript<'a>,
-//     size:       Option<u64>,
-//     shard_size: Option<u64>,
-//     order:      Option<Order<'a>>
-// }
-
-// impl<'a> Terms<'a> {
-//     pub fn new<FOS: Into<FieldOrScript<'a>>>(field: FOS) -> Terms<'a> {
-//         Terms {
-//             field:      field.into(),
-//             size:       None,
-//             shard_size: None,
-//             order:      None
-//         }
-//     }
-
-//     add_field!(with_size, size, u64);
-//     add_field!(with_shard_size, shard_size, u64);
-//     add_field!(with_order, order, Order<'a>);
-// }
-
-// impl<'a> ToJson for Terms<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut json = BTreeMap::new();
-//         self.field.add_to_object(&mut json);
-
-//         optional_add!(self, json, size);
-//         optional_add!(self, json, shard_size);
-//         optional_add!(self, json, order);
-
-//         Json::Object(json)
-//     }
-// }
-
-// bucket_agg!(Terms);
-
-// Range aggs and dependencies
-
-/// A specific range, there will be many of these making up a range aggregation
-#[derive(Debug)]
-pub struct RangeInst<'a> {
-    from: Option<JsonVal>,
-    to:   Option<JsonVal>,
-    key:  Option<&'a str>
-}
-
-impl<'a> RangeInst<'a> {
-    pub fn new() -> RangeInst<'a> {
-        RangeInst {
-            from: None,
-            to:   None,
-            key:  None
-        }
-    }
-
-    add_field!(with_from, from, JsonVal);
-    add_field!(with_to, to, JsonVal);
-    add_field!(with_key, key, &'a str);
-}
-
-impl<'a> ToJson for RangeInst<'a> {
-    fn to_json(&self) -> Json {
-        let mut d = BTreeMap::new();
-
-        optional_add!(self, d, from);
-        optional_add!(self, d, to);
-        optional_add!(self, d, key);
-
-        Json::Object(d)
-    }
-}
-
-// /// Range aggregations
-// ///
-// /// The keyed option will always be used.
-// ///
-// /// https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-range-aggregation.html
-// #[derive(Debug)]
-// pub struct Range<'a> {
-//     field: FieldOrScript<'a>,
-//     keyed: bool,
-//     ranges: Vec<RangeInst<'a>>
-// }
-
-// impl<'a> Range<'a> {
-//     pub fn new<FOS: Into<FieldOrScript<'a>>>(field: FOS,
-//                                              ranges: Vec<RangeInst<'a>>) -> Range<'a> {
-//         Range {
-//             field:  field.into(),
-//             keyed:  true,
-//             ranges: ranges
-//         }
-//     }
-
-//     pub fn inst() -> RangeInst<'a> {
-//         RangeInst::new()
-//     }
-// }
-
-// impl<'a> ToJson for Range<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut json = BTreeMap::new();
-//         self.field.add_to_object(&mut json);
-//         json.insert("keyed".to_owned(), Json::Boolean(self.keyed));
-//         json.insert("ranges".to_owned(), self.ranges.to_json());
-//         Json::Object(json)
-//     }
-// }
-
-// bucket_agg!(Range);
-
-/// A specific element of a range for a `DateRange` aggregation
-#[derive(Debug)]
-pub struct DateRangeInst<'a> {
-    from: Option<&'a str>,
-    to:   Option<&'a str>
-}
-
-impl<'a> DateRangeInst<'a> {
-    pub fn new() -> DateRangeInst<'a> {
-        DateRangeInst {
-            from: None,
-            to:   None
-        }
-    }
-
-    add_field!(with_from, from, &'a str);
-    add_field!(with_to, to, &'a str);
-}
-
-impl<'a> ToJson for DateRangeInst<'a> {
-    fn to_json(&self) -> Json {
-        let mut d = BTreeMap::new();
-        optional_add!(self, d, from);
-        optional_add!(self, d, to);
-
-        Json::Object(d)
-    }
-}
-
-// /// Date range aggregation.  See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-daterange-aggregation.html
-// #[derive(Debug)]
-// pub struct DateRange<'a> {
-//     field: FieldOrScript<'a>,
-//     format: Option<&'a str>,
-//     ranges: Vec<DateRangeInst<'a>>
-// }
-
-// impl<'a> DateRange<'a> {
-//     pub fn new<FOS: Into<FieldOrScript<'a>>>(field: FOS,
-//                                              ranges: Vec<DateRangeInst<'a>>) -> DateRange<'a> {
-//         DateRange {
-//             field: field.into(),
-//             format: None,
-//             ranges: ranges
-//         }
-//     }
-
-//     pub fn inst() -> DateRangeInst<'a> {
-//         DateRangeInst::new()
-//     }
-// }
-
-// impl<'a> ToJson for DateRange<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut json = BTreeMap::new();
-//         self.field.add_to_object(&mut json);
-//         optional_add!(self, json, format);
-//         json.insert("ranges".to_owned(), self.ranges.to_json());
-//         Json::Object(json)
-//     }
-// }
-
-// bucket_agg!(DateRange);
-
-/// Histogram aggregation.
-#[derive(Debug)]
-pub struct ExtendedBounds {
-    min: i64,
-    max: i64
-}
-
-impl ExtendedBounds {
-    pub fn new(min: i64, max: i64) -> ExtendedBounds {
-        ExtendedBounds {
-            min: min,
-            max: max
-        }
-    }
-}
-
-impl ToJson for ExtendedBounds {
-    fn to_json(&self) -> Json {
-        let mut d = BTreeMap::new();
-        d.insert("min".to_owned(), Json::I64(self.min));
-        d.insert("max".to_owned(), Json::I64(self.max));
-
-        Json::Object(d)
-    }
-}
-
-impl From<(i64, i64)> for ExtendedBounds {
-    fn from(from: (i64, i64)) -> ExtendedBounds {
-        ExtendedBounds::new(from.0, from.1)
-    }
-}
-
-// #[derive(Debug)]
-// pub struct Histogram<'a> {
-//     field:           &'a str,
-//     interval:        Option<u64>,
-//     min_doc_count:   Option<u64>,
-//     extended_bounds: Option<ExtendedBounds>,
-//     order:           Option<Order<'a>>
-// }
-
-// impl<'a> Histogram<'a> {
-//     pub fn new(field: &'a str) -> Histogram<'a> {
-//         Histogram {
-//             field: field,
-//             interval: None,
-//             min_doc_count: None,
-//             extended_bounds: None,
-//             order: None
-//         }
-//     }
-
-//     add_field!(with_interval, interval, u64);
-//     add_field!(with_min_doc_count, min_doc_count, u64);
-//     add_field!(with_extended_bounds, extended_bounds, ExtendedBounds);
-//     add_field!(with_order, order, Order<'a>);
-// }
-
-// impl<'a> ToJson for Histogram<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         d.insert("field".to_owned(), self.field.to_json());
-//         optional_add!(self, d, interval);
-//         optional_add!(self, d, min_doc_count);
-//         optional_add!(self, d, extended_bounds);
-//         optional_add!(self, d, order);
-
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(Histogram);
-
-/// Date histogram and related fields
-#[derive(Debug)]
-pub enum TimeZone<'a> {
-    Offset(u64),
-    Str(&'a str)
-}
-
-impl<'a> ToJson for TimeZone<'a> {
-    fn to_json(&self) -> Json {
-        match self {
-            &TimeZone::Offset(offset) => Json::U64(offset),
-            &TimeZone::Str(tz_str)    => tz_str.to_json()
-        }
-    }
-}
-
-impl<'a> From<&'a str> for TimeZone<'a> {
-    fn from(from: &'a str) -> TimeZone<'a> {
-        TimeZone::Str(from)
-    }
-}
-
-impl<'a> From<u64> for TimeZone<'a> {
-    fn from(from: u64) -> TimeZone<'a> {
-        TimeZone::Offset(from)
-    }
-}
-
-#[derive(Debug)]
-pub enum Interval {
-    Year,
-    Quarter,
-    Month,
-    Week,
-    Day,
-    Hour,
-    Minute,
-    Second
-}
-
-impl ToJson for Interval {
-    fn to_json(&self) -> Json {
-        use self::Interval::*;
-        match *self {
-            Year => "year",
-            Quarter => "quarter",
-            Month => "month",
-            Week => "week",
-            Day => "day",
-            Hour => "hour",
-            Minute => "minute",
-            Second => "second"
-        }.to_json()
-    }
-}
-
-// #[derive(Debug)]
-// pub struct DateHistogram<'a> {
-//     field: &'a str,
-//     interval: Interval,
-//     time_zone: Option<TimeZone<'a>>,
-//     offset: Option<Duration>,
-//     format: Option<&'a str>,
-// }
-
-// impl<'a> DateHistogram<'a> {
-//     pub fn new<I>(field: &'a str, interval: I) -> DateHistogram<'a>
-//     where I: Into<Interval> {
-//         DateHistogram {
-//             field: field,
-//             interval: interval.into(),
-//             time_zone: None,
-//             offset: None,
-//             format: None
-//         }
-//     }
-
-//     add_field!(with_time_zone, time_zone, TimeZone<'a>);
-//     add_field!(with_offset, offset, Duration);
-//     add_field!(with_format, format, &'a str);
-// }
-
-// impl<'a> ToJson for DateHistogram<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         d.insert("field".to_owned(), self.field.to_json());
-//         d.insert("interval".to_owned(), self.interval.to_json());
-//         optional_add!(self, d, time_zone);
-//         optional_add!(self, d, offset);
-//         optional_add!(self, d, format);
-
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(DateHistogram);
-
-// #[derive(Debug)]
-// pub struct GeoDistanceInst {
-//     from: Option<f64>,
-//     to:   Option<f64>
-// }
-
-// impl GeoDistanceInst {
-//     pub fn new() -> GeoDistanceInst {
-//         GeoDistanceInst {
-//             from: None,
-//             to:   None
-//         }
-//     }
-
-//     add_field!(with_from, from, f64);
-//     add_field!(with_to, to, f64);
-// }
-
-// impl ToJson for GeoDistanceInst {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-//         optional_add!(self, d, from);
-//         optional_add!(self, d, to);
-
-//         Json::Object(d)
-//     }
-// }
-
-// #[derive(Debug)]
-// pub struct GeoDistance<'a> {
-//     field:         &'a str,
-//     origin:        &'a Location,
-//     unit:          Option<DistanceUnit>,
-//     distance_type: Option<DistanceType>,
-//     ranges:        &'a [GeoDistanceInst]
-// }
-
-// impl<'a> GeoDistance<'a> {
-//     pub fn new(field: &'a str,
-//                origin: &'a Location,
-//                ranges: &'a [GeoDistanceInst]) -> GeoDistance<'a> {
-//         GeoDistance {
-//             field:         field,
-//             origin:        origin,
-//             unit:          None,
-//             distance_type: None,
-//             ranges:        ranges
-//         }
-//     }
-
-//     add_field!(with_unit, unit, DistanceUnit);
-//     add_field!(with_distance_type, distance_type, DistanceType);
-
-//     pub fn inst() -> GeoDistanceInst {
-//         GeoDistanceInst::new()
-//     }
-// }
-
-// impl<'a> ToJson for GeoDistance<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut json = BTreeMap::new();
-
-//         json.insert("field".to_owned(), self.field.to_json());
-//         json.insert("origin".to_owned(), self.origin.to_json());
-//         json.insert("ranges".to_owned(), self.ranges.to_json());
-
-//         optional_add!(self, json, unit);
-//         optional_add!(self, json, distance_type);
-
-//         Json::Object(json)
-//     }
-// }
-
-// bucket_agg!(GeoDistance);
-
-// /// Geohash aggregation
-// #[derive(Debug)]
-// pub struct GeoHash<'a> {
-//     field:      &'a str,
-//     precision:  Option<u64>,
-//     size:       Option<u64>,
-//     shard_size: Option<u64>
-// }
-
-// impl<'a> GeoHash<'a> {
-//     pub fn new(field: &'a str) -> GeoHash<'a> {
-//         GeoHash {
-//             field: field,
-//             precision: None,
-//             size: None,
-//             shard_size: None
-//         }
-//     }
-
-//     add_field!(with_precision, precision, u64);
-//     add_field!(with_size, size, u64);
-//     add_field!(with_shard_size, shard_size, u64);
-// }
-
-// impl<'a> ToJson for GeoHash<'a> {
-//     fn to_json(&self) -> Json {
-//         let mut d = BTreeMap::new();
-
-//         d.insert("field".to_owned(), self.field.to_json());
-
-//         optional_add!(self, d, precision);
-//         optional_add!(self, d, size);
-//         optional_add!(self, d, shard_size);
-
-//         Json::Object(d)
-//     }
-// }
-
-// bucket_agg!(GeoHash);
-
-/// The set of bucket aggregations
-#[derive(Debug)]
-pub enum BucketAggregation<'a> {
-    Global(Global<'a>),
-    // Filter(Filter<'a>),
-    // Filters(Filters<'a>),
-    // Missing(Missing<'a>),
-    // Nested(Nested<'a>),
-    // ReverseNested(ReverseNested<'a>),
-    // Children(Children<'a>),
-    // Terms(Terms<'a>),
-    // Range(Range<'a>),
-    // DateRange(DateRange<'a>),
-    // Histogram(Histogram<'a>),
-    // DateHistogram(DateHistogram<'a>),
-    // GeoDistance(GeoDistance<'a>),
-    // GeoHash(GeoHash<'a>)
-}
-
-impl<'a> BucketAggregation<'a> {
-    fn add_to_object<S>(&self, json: &mut BTreeMap<&'static str, Value>)
-        where S: Serialize {
-
-        let (key, value) = match self {
-            &BucketAggregation::Global(ref a) => ("global", to_value(a))
-        };
-
-        json.insert(key, value);
-    }
-}
-
 /// Aggregations are either metrics or bucket-based aggregations
 #[derive(Debug)]
 pub enum Aggregation<'a> {
     /// A metric aggregation (e.g. min)
-    Metrics(MetricsAggregation<'a>),
+    Metrics(metrics::MetricsAggregation<'a>),
 
     /// A bucket aggregation, groups data into buckets and optionally applies
     /// sub-aggregations
-    Bucket(BucketAggregation<'a>, Option<Aggregations<'a>>)
+    Bucket(bucket::BucketAggregation<'a>, Option<Aggregations<'a>>)
 }
 
 impl<'a> Serialize for Aggregation<'a> {
@@ -1342,7 +245,8 @@ impl<'a> Aggregations<'a> {
     /// # Examples
     ///
     /// ```
-    /// use rs_es::operations::search::aggregations::{Aggregations, Min};
+    /// use rs_es::operations::search::aggregations::Aggregations;
+    /// use rs_es::operations::search::aggregations::metrics::Min;
     ///
     /// let mut aggs = Aggregations::new();
     /// aggs.add("agg_name", Min::new("field_name"));
